# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MarocExplore API",
        "description": "API documentation for the MarocExplore project, providing endpoints for itinerary management, destinations, and user authentication.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://marocexplore.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/itineraries": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "List all itineraries",
                "operationId": "f662c752b9ff94a452943ba63b6606b8",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by itinerary title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "Search by destination name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Create a new itinerary",
                "operationId": "ffd76e51a0b5db2f1ec3f04e62866d98",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "image",
                                    "destinations"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Summer Trip"
                                    },
                                    "category": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "destinations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "type": "string"
                                                },
                                                "address": {
                                                    "type": "string"
                                                },
                                                "places": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "activities": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "dishes": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Itinerary created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itineraries/{itinerary}": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Get a single itinerary details",
                "operationId": "79238560f34cff7e78f41d8721b37850",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Itinerary not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Update an itinerary",
                "description": "Uses POST with _method=PUT for multipart/form-data compatibility",
                "operationId": "d6914ceca7ba7ea5864cc7f02c42b13e",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "category": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "visiting",
                                            "visited",
                                            "canceled"
                                        ]
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "removed_activities": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "removed_places": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "removed_dishes": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "removed_destinations": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "destinations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "title": {
                                                    "type": "string"
                                                },
                                                "address": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Itinerary updated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Delete an itinerary",
                "operationId": "63e96427d9843d8e5ed3bf3b8dbfc412",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itinerary deleted"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itineraries/my": {
            "get": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Get itineraries of the authenticated user",
                "operationId": "02b39d03262babaef5165d3a3563a8f9",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/itineraries/{itinerary}/copy": {
            "post": {
                "tags": [
                    "Itineraries"
                ],
                "summary": "Copy an itinerary",
                "operationId": "a1d5a4851b2cc29f35867291c2a77ce7",
                "parameters": [
                    {
                        "name": "itinerary",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Itinerary copied successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}